requirements.txt
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
distro==1.9.0
exceptiongroup==1.2.2
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2